model.py
import torch
from transformers import AutoTokenizer, AutoModelForCausalLM
from llama_cpp import Llama

def load_model():
    # Load the model using llama.cpp
    # This provides an optimized inference for the LLaMA model
    model = Llama(model_path="path/to/ggml/model.bin", n_ctx=2048, n_threads=4)
    return model

def generate_text(model, prompts, max_lengths):
    results = []
    for prompt, max_length in zip(prompts, max_lengths):
        # Generate text using the llama.cpp model
        output = model(prompt, max_tokens=max_length)
        results.append(output['choices'][0]['text'])
    return results