requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
torch==1.9.0
transformers==4.11.3
llama-cpp-python==0.1.50
pytest==6.2.5
aiohttp==3.7.4
pydantic==1.8.2
numpy==1.21.2
tqdm==4.62.3